Move test_qtbot_wait to test_basics and stop_watch fixture to conftest
