+++ tests/test_basics.py
@@ -204,6 +204,13 @@ def test_widgets_closed_before_fixtures(testdir):
     ])
 
 
+def test_qtbot_wait(qtbot, stop_watch):
+    stop_watch.start()
+    qtbot.wait(250)
+    stop_watch.stop()
+    assert stop_watch.elapsed >= 220
+
+
 class EventRecorder(QWidget):
 
     """
